rename: 가독성을 고려한 클래스명 정정
